Align with ROS 2 Dashing
We have the latest stable ROS 2 - Dashing released recently. To make our
Node.js client workable on it, this PR makes a series of changes to achieve
this goal, please see below:

- Support wstring
(https://design.ros2.org/articles/idl_interface_definition.html)
- Support default value
- API updates for rmw preallocation work
- Update the test case, as some message has been removed or changed

Fix #493
