+++ rosidl_gen/generator.json
@@ -1,6 +1,6 @@
 {
   "name": "rosidl-generator",
-  "version": "0.2.1",
+  "version": "0.3.0",
   "description": "Generate JavaScript object from ROS IDL(.msg) files",
   "main": "index.js",
   "authors": [
